requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
Brotli==1.1.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.4
cssselect2==0.8.0
Django==5.1.6
django-allauth==65.9.0
django-cors-headers==4.7.0
django-countries==7.6.1
django-crispy-forms==2.4
django-extensions==4.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
et_xmlfile==2.0.0
fonttools==4.59.1
html5lib==1.1
idna==3.10
jwt==1.3.1
lxml==6.0.1
node==1.2.2
npm==0.1.1
odict==1.9.0
openpyxl==3.1.5
optional-django==0.3.0
oscrypto==1.3.0
pillow==11.1.0
plumber==1.7
psycopg2-binary==2.9.10
pycparser==2.22
pydyf==0.11.0
pyHanko==0.30.0
pyhanko-certvalidator==0.28.0
PyJWT==2.10.1
pypdf==6.0.0
pyphen==0.17.2
python-bidi==0.6.6
PyYAML==6.0.2
react==4.3.0
reportlab==4.4.3
requests==2.32.3
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
stripe==12.2.0
svglib==1.5.1
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.14.0
tzdata==2025.1
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.4.0
weasyprint==66.0
webencodings==0.5.1
xhtml2pdf==0.2.17
zope.component==6.0
zope.deferredimport==5.0
zope.deprecation==5.1
zope.event==5.0
zope.hookable==7.0
zope.interface==7.2
zope.lifecycleevent==5.1
zope.proxy==6.1
zopfli==0.2.3.post1
